# contracts/openapi/inventory.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventory API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "operationId": "listProducts",
        "responses": {
          "200": {
            "description": "Productos ordenados por identificador.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductStock"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "sku": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "initialStock": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000000
                  }
                },
                "required": [
                  "sku",
                  "name",
                  "initialStock"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resultado persistido; replay histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "productId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "sku": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    },
                    "onHand": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "reserved": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "available": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "productId",
                    "sku",
                    "name",
                    "onHand",
                    "reserved",
                    "available",
                    "version",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Problem Details: 400 inválido, 404 inexistente, 409 conflicto, 503 transitorio con Retry-After.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "recountStock",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "productId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "stock": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000000
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "productId",
                  "stock",
                  "expectedVersion",
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado persistido; replay histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "productId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "movementId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "quantity": {
                      "type": "integer",
                      "minimum": -1000000000,
                      "maximum": 1000000000
                    },
                    "previousOnHand": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "onHand": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "reserved": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "available": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    }
                  },
                  "required": [
                    "productId",
                    "movementId",
                    "quantity",
                    "previousOnHand",
                    "onHand",
                    "reserved",
                    "available",
                    "version",
                    "reason"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Problem Details: 400 inválido, 404 inexistente, 409 conflicto, 503 transitorio con Retry-After.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/products/{productId}/restock": {
      "post": {
        "operationId": "restockProduct",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "movementId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000000000
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "movementId",
                  "quantity",
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resultado persistido; replay histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "productId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "movementId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "quantity": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 1000000000
                    },
                    "previousOnHand": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "onHand": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "reserved": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "available": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    }
                  },
                  "required": [
                    "productId",
                    "movementId",
                    "quantity",
                    "previousOnHand",
                    "onHand",
                    "reserved",
                    "available",
                    "version",
                    "reason"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Problem Details: 400 inválido, 404 inexistente, 409 conflicto, 503 transitorio con Retry-After.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/products/{productId}/stock": {
      "get": {
        "operationId": "getStock",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado persistido; replay histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "productId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "onHand": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "reserved": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "available": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1000000000
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "productId",
                    "onHand",
                    "reserved",
                    "available",
                    "version",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Problem Details: 400 inválido, 404 inexistente, 409 conflicto, 503 transitorio con Retry-After.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductStock": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "productId": { "type": "string", "format": "uuid" },
          "sku": { "type": "string" },
          "name": { "type": "string" },
          "onHand": { "type": "integer", "minimum": 0, "maximum": 1000000000 },
          "reserved": { "type": "integer", "minimum": 0, "maximum": 1000000000 },
          "available": { "type": "integer", "minimum": 0, "maximum": 1000000000 },
          "version": { "type": "integer", "minimum": 1 },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["productId", "sku", "name", "onHand", "reserved", "available", "version", "updatedAt"]
      }
    }
  }
}